\name{A2Rplot}
\alias{A2Rplot}
\title{Colored dendrogram}
\source{
  \url{http://addictedtor.free.fr/packages/A2R/lastVersion/}
}
\usage{
  A2Rplot(x, ...)
}
\description{
  This function plot an hclust dendrogram with one color
  per branch (see examples).
}
\note{
  The A2R package has not been updated since January 2006
  and cannot be installed anymore with a recent version of
  R. It's why this function has been copied here.
}
\author{
  Romain Francois
}
\seealso{
  \url{http://rpubs.com/gaston/dendrograms}
}

