% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_index_group_rows.R
\name{auto_index_group_rows}
\alias{auto_index_group_rows}
\title{Compute automatically the index for grouping rows with kableExtra}
\usage{
auto_index_group_rows(x)
}
\arguments{
\item{x}{a vector}
}
\description{
When using \code{group_rows} function from \code{kableExtra}, you can provide a names vector providing
the number of rows to group together. \code{auto_index_group_rows} compute that index based on a vector
by counting the successive identical values of that vector.
}
\examples{
x <- c("a", "a", "a", "b", "b", "a", "c", "c")
auto_index_group_rows(x)
}
