\name{best.cutree}
\alias{best.cutree}
\title{Optimal partition based on the higher relative loss of inertia}
\source{
  \url{http://joseph.larmarange.net/?Ou-couper-un-dendrogramme}
}
\usage{
  best.cutree(hc, min = 3, max = 20, loss = FALSE,
    graph = FALSE, ...)
}
\description{
  Optimal partition based on the higher relative loss of
  inertia
}

