\name{cutree.order}
\alias{cutree.order}
\title{Cut a tree into groups of data in the order of the tree}
\source{
  \url{http://addictedtor.free.fr/packages/A2R/lastVersion/}
}
\usage{
  cutree.order(hclu, k = NULL, h = NULL)
}
\description{
  Cut a tree (result from \link{hclust}) into groups of
  data. Groups are in the order of the tree leafs.
}
\note{
  The A2R package has not been updated since January 2006
  and cannot be installed anymore with a recent version of
  R. It's why this function has been copied here.
}
\author{
  Romain Francois
}

