% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formating.R
\name{en}
\alias{en}
\alias{en0}
\alias{en1}
\alias{en2}
\alias{en3}
\alias{en4}
\alias{en5}
\title{English formatter}
\usage{
en(x, digits = 1)

en0(x)

en1(x)

en2(x)

en3(x)

en4(x)

en5(x)
}
\arguments{
\item{x}{a numeric vector to format}

\item{digits}{number of decimal digits to display}
}
\description{
Use a dot as decimal separator and a single quote as separator for thousands.
}
\details{
\code{en0(x)} is a shortcut for \code{en(x, digits = 0)},
\code{en1(x)} for \code{en(x, digits = 1)}, etc.
}
\examples{
en(12345.67)
en0(c(0.123, 123.4567, 1234567.89))
en2(c(0.123, 123.4567, 1234567.89))
en5(c(0.123, 123.4567, 1234567.89))
}

