% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formating.R
\name{f}
\alias{f}
\alias{f0}
\alias{f1}
\alias{f2}
\alias{f3}
\alias{f4}
\alias{f5}
\title{French formatter}
\usage{
f(x, digits = 1)

f0(x)

f1(x)

f2(x)

f3(x)

f4(x)

f5(x)
}
\arguments{
\item{x}{a numeric vector to format}

\item{digits}{number of decimal digits to display}
}
\description{
Use a comma as decimal separator and a space as separator for thousands (French way).
}
\details{
\code{f0(x)} is a shortcut for \code{f(x, digits = 0)},
\code{f1(x)} for \code{f(x, digits = 1)}, etc.
}
\examples{
f(0.004)
f0(c(0.000045, 0.56789, 234.567))
f2(c(0.000045, 0.56789, 234.567))
f5(c(0.000045, 0.56789, 234.567))
}

