% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formating.R
\name{fr}
\alias{fr}
\alias{fr0}
\alias{fr1}
\alias{fr2}
\alias{fr3}
\alias{fr4}
\alias{fr5}
\title{French formatter}
\usage{
fr(x, digits = 1)

fr0(x)

fr1(x)

fr2(x)

fr3(x)

fr4(x)

fr5(x)
}
\arguments{
\item{x}{a numeric vector to format}

\item{digits}{number of decimal digits to display}
}
\description{
Use a comma as decimal separator and a space as separator for thousands (French way).
}
\details{
\code{fr0(x)} is a shortcut for \code{fr(x, digits = 0)},
\code{fr1(x)} for \code{fr(x, digits = 1)}, etc.
}
\examples{
fr(0.004)
fr0(c(0.000045, 0.56789, 234.567))
fr2(c(0.000045, 0.56789, 234.567))
fr5(c(0.000045, 0.56789, 234.567))
}

