% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legend.R
\name{get_legend}
\alias{get_legend}
\title{Extract the legend of a ggplot2}
\usage{
get_legend(p)
}
\arguments{
\item{p}{a ggplot with a legend}
}
\description{
Extract the legend of a ggplot2
}
\note{
Could be combined with \code{\link{multiplot}} (see examples).
}
\examples{
require(ggplot2)
p1 <- ggplot(iris) +
  aes(x = Sepal.Length, y = Sepal.Width, color = Species) +
  geom_point() +
  theme(legend.position = "bottom")
p2 <- ggplot(iris) +
  aes(x = Petal.Length, y = Petal.Width, color = Species) +
  geom_point()

# get legend from p1
l <- get_legend(p1)

# remove legends
p1 <- p1 + theme(legend.position = "none")
p2 <- p2 + theme(legend.position = "none")

# Combining
multiplot(p1, p2, l, heights = c(2, 2, 1))
}
\references{
Adapted from \url{http://stackoverflow.com/questions/12539348/ggplot-separate-legend-and-plot}
}
\seealso{
\url{https://github.com/hadley/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs}
}
