% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg909090.R
\name{gg909090}
\alias{gg909090}
\title{plot the 90-90-90 UNAIDS's target}
\usage{
gg909090(first = 0.9, second = 0.9, third = 0.9, col = "#E81634",
  sub = c("diagnosed", "on treatment", "virally suppressed"), p_size = 6,
  sub_size = 4)
}
\arguments{
\item{first}{value of the first 90}

\item{second}{value of the second 90}

\item{third}{value of the third 90}

\item{col}{colour of the graph}

\item{sub}{value of the sub-labels (\code{NULL} to hide)}

\item{p_size}{text size of the percentages}

\item{sub_size}{text size of the sub-labels}
}
\description{
plot the 90-90-90 UNAIDS's target
}
\examples{
gg909090()
}

