% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcross.R
\name{ggcross}
\alias{ggcross}
\title{Cross-tabulation matrix plot}
\usage{
ggcross(
  formula,
  data,
  weight = NULL,
  fill = "stdres",
  fill_breaks = c(-3, -2, 2, 3),
  fill_title = "Standardized residuals",
  size = "observed",
  max_size = 20,
  size_title = "Observations",
  labels = NULL,
  labels_size = 3.5,
  svystatistic = c("F", "Chisq", "Wald", "adjWald", "lincom", "saddlepoint"),
  return_data = FALSE
)
}
\arguments{
\item{formula}{formula of variables to be cross-tabulated,
rows on left hand side and columns on the right hand side}

\item{data}{data frame or survey object containing the data}

\item{weight}{optionnal string indicating a column containing weights}

\item{fill}{variable mapped to fill OR color name OR \code{NULL} (see examples)}

\item{fill_breaks}{how to cut fill variable into categories? (cf. \code{\link[base]{cut}})}

\item{fill_title}{legend title for fill}

\item{size}{variable mapped to size OR \code{"raster"} OR \code{NULL} (see examples)}

\item{max_size}{size of largest point}

\item{size_title}{legend title for size}

\item{labels}{optionnal cell labels (see examples)}

\item{labels_size}{size of cell labels}

\item{svystatistic}{statistic to be used for survey data (see \code{\link[survey]{svychisq}})}

\item{return_data}{return computed data.frame instead of plot?}
}
\value{
a ggplot graphic or a data frame if \code{return_data == TRUE}.
}
\description{
\lifecycle{deprecated}

For a more recent and powerful function, use \code{GGally::ggtable()}
}
\details{
By default, the size of squares represents the number of observations while the color
represents the chi-squared residuals.
}
\examples{
ggcross(Sex + Age + Class ~ Survived, data = as.data.frame(Titanic), weight = "Freq")
ggcross(Sex + Age + Class ~ Survived,
  data = as.data.frame(Titanic), weight = "Freq",
  fill_breaks = c(-3, -2, 0, 2, 3)
)
ggcross(Sex + Age + Class ~ Survived,
  data = as.data.frame(Titanic), weight = "Freq",
  fill = "blue"
)
ggcross(Sex + Age + Class ~ Survived,
  data = as.data.frame(Titanic), weight = "Freq",
  size = "expected", size_title = "Expected number of observations"
)
ggcross(Sex + Age + Class ~ Survived,
  data = as.data.frame(Titanic), weight = "Freq",
  labels = "scales::percent(row.prop)", size = NULL
)
ggcross(Sex + Age + Class ~ Survived,
  data = as.data.frame(Titanic), weight = "Freq",
  labels = "scales::percent(row.prop)", size = "raster"
)
ggcross(Sex + Age + Class ~ Survived,
  data = as.data.frame(Titanic), weight = "Freq",
  return_data = TRUE
)
if (require(survey)) {
  data(api)
  dclus1 <- svydesign(id = ~dnum, weights = ~pw, data = apiclus1, fpc = ~fpc)
  ggcross(awards ~ cname, data = dclus1)
}
}
