% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concatenate.R
\name{\%p\%}
\alias{\%p\%}
\title{Concatenate operator}
\usage{
x \%p\% y
}
\arguments{
\item{x}{a first vector to paste}

\item{y}{a second vector to paste}
}
\description{
The \code{\%p\%} operator is a shortcut for \code{\link[base]{paste0}}.
}
\examples{
"abc" \%p\% "def"
"v_" \%p\% 1:3
c("A", "B", "C") \%p\% 1:3
}

