% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{int}
\alias{int}
\alias{int0}
\alias{int1}
\alias{int2}
\alias{int3}
\alias{int4}
\alias{int5}
\title{International formatter}
\usage{
int(x, digits = 1)

int0(x)

int1(x)

int2(x)

int3(x)

int4(x)

int5(x)
}
\arguments{
\item{x}{a numeric vector to format}

\item{digits}{number of decimal digits to display}
}
\description{
Use a dot as decimal separator and a space as separator for thousands.
}
\details{
\code{int0(x)} is a shortcut for \code{int(x, digits = 0)}, 
\code{int1(x)} for \code{int(x, digits = 1)}, etc.
}
\examples{
int(12345.67)
int0(c(0.123, 123.4567, 1234567.89))
int2(c(0.123, 123.4567, 1234567.89))
int5(c(0.123, 123.4567, 1234567.89))
}

