% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_cumsum.R
\name{inv_cumsum}
\alias{inv_cumsum}
\title{Inverted cumulative sum}
\usage{
inv_cumsum(x)
}
\arguments{
\item{x}{numerical vector}
}
\description{
This is a decreasing version of \code{cumsum}.
}
\examples{
cumsum(1:10)
inv_cumsum(1:10)
}

