\name{JLutils-package}
\alias{JLutils-package}
\alias{JLutils}
\docType{package}
\title{
JLutils
}
\description{
A collection of diverse functions who don't fit (yet?) in another package.
}
\source{\url{https://github.com/larmarange/JLutils}}
\author{Joseph Larmarange - \url{http://joseph.larmarange.net}}
\keyword{ package }
