% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/label_freq.R
\name{lfactor}
\alias{lfactor}
\title{Labeled factor}
\usage{
lfactor(labeled_x, labeled_df = NULL, add.codes = TRUE)
}
\arguments{
\item{labeled_x}{the name of a column or a vector}

\item{labeled_df}{a data frame (if \code{labeled_x} is the column name)}

\item{add.codes}{add the numerical code to the labels?}
}
\value{
If \code{labeled_x} is already a factor, \code{labeled_x} is
returned with no change.
}
\description{
Convert to a factor using SPSS/Stata value labels if they are available
in the data frame
}
\details{
This function is particulary useful when the data has been imported
from SPSS or Stata but the labeled variables not converted to factors.

It's possible to enter \code{lfactor(df$Sex)} for convenient purpose. In this
situation, the function will try to guess what is the corresponding data frame
(i.e. \code{df}) by analysing the function call. To avoid any issue when
programming, you should prefer \code{lfactor("Sex", df)}.
}
\examples{
\dontrun{
   lfactor("Sex", df)
   lfactor(df$Sex)
}
}
\seealso{
\code{\link{ltabs}}, \code{\link{lfreq}}
}

