% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/label_freq.R
\name{lfreq}
\alias{lfreq}
\title{Frequency table with value labels}
\usage{
lfreq(x, ...)
}
\arguments{
\item{x}{a vector}

\item{...}{additional arguments sent to \code{\link[questionr]{freq}}}
}
\description{
This function display a frequency table with value labels if they are available
}
\details{
This function is particulary useful when the data has been imported
from SPSS or Stata but the labelled variables not converted to factors.
}
\note{
This function is an equivalent of applying \code{\link[questionr]{freq}} on
the result of \code{\link{ltabs}}.
}
\examples{
\dontrun{
   lfreq(df$Sex)
}
}
\seealso{
\code{\link{ltabs}}
}

