% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_right.R
\name{lm_right}
\alias{lm_right}
\alias{lm_left}
\alias{lm_zero}
\title{Extend geom_smooth in a single direction}
\usage{
lm_right(formula, data, ...)

lm_left(formula, data, ...)

lm_zero(formula, data, ...)
}
\arguments{
\item{formula}{a model formula}

\item{data}{a data.frame}

\item{...}{unused}
}
\description{
Allows to extend a linear regression line drawn with \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}
to only one direction (left or right), or to zero. See examples.
DOES NOT WORK WITH CURRENT VERSION OF GGPLOT2.
}
\details{
cf. https://github.com/tidyverse/ggplot2/issues/3132
The implementation should be through a new stat rather than extending
\code{\link[=predictdf]{predictdf()}}
}
\note{
Adapted from \url{https://stackoverflow.com/questions/26705554/extend-geom-smooth-in-a-single-direction}
}
\examples{
p <- ggplot(iris) + 
  aes(x = Petal.Width, y = Petal.Length, colour = Species) +
  geom_point()
  
# classic geom_smooth
p + geom_smooth(method = "lm")

# use `fullrange = TRUE` to extend on both sides
p + geom_smooth(method = "lm", fullrange = TRUE)

# use `method = "lm_right"` or `method = "lm_left"`
# with `fullrange = TRUE` to extand in only one direction
p + geom_smooth(method = "lm_right", fullrange = TRUE)
p + geom_smooth(method = "lm_left", fullrange = TRUE)

# use `method = "lm_zero"` to include zero
p + geom_smooth(method = "lm_zero")
}
