% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_to_periods.R
\name{long_to_periods}
\alias{long_to_periods}
\title{Transform data from long format to period format}
\usage{
long_to_periods(.data, .id, .time, .status = NULL, .by = NULL)
}
\arguments{
\item{.data}{a data.frame}

\item{.id}{a character, column containing individual ids}

\item{.time}{a character, column containing time variable}

\item{.status}{a character vector, a status variable (optionnal)}

\item{.by}{a character vector, co-variables to consider (optionnal)}
}
\description{
Transform data from long format to period format
}
\examples{
\dontrun{
load(url("https://larmarange.github.io/analyse-R/data/care_trajectories.RData"))
care_trajectories
long_to_periods(care_trajectories, .id = "id", .time = "month")
long_to_periods(care_trajectories, .id = "id", .time = "month", .by = c("sex", "age", "care_status"))
long_to_periods(care_trajectories, .id = "id", .time = "month", .by = c("sex", "age"), .status = ("care_status"))
}
}
