% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/label_freq.R
\name{ltabs}
\alias{ltabs}
\title{Cross tabulation with SPSS/Stata value labels}
\usage{
ltabs(labeled_x, labeled_df = NULL, add.variable.labels = FALSE, ...)
}
\arguments{
\item{labeled_x}{a formula (see \code{\link[stats]{xtabs}}) or a vector}

\item{add.variable.labels}{add variable labels to dimension names?}

\item{...}{additional parameters sent to \code{\link[stats]{xtabs}}}

\item{labeled_d}{a data.frame (if \code{labeled_x} is a formula) or an optional second vector}
}
\description{
Create a contingency table (with \code{\link[stats]{xtabs}}) and add value and variable
labels if they are available (typically if the data frame has been imported with
\code{\link[foreign]{read.spss}} or \code{\link[foreign]{read.dta}})
}
\details{
This function is particulary useful when the data has been imported
from SPSS or Stata but the labeled variables not converted to factors.

It's possible to enter \code{ltabs(df$Sex)} for convenient purpose. In this
situation, the function will try to guess what is the corresponding data frame
(i.e. \code{df}) by analysing the function call. To avoid any issue when
programming, you should prefer \code{ltabs(~Sex, df)}.
}
\examples{
\dontrun{
   ltabs(~Sex, df)
   ltabs(~Sex + Age, df)
   ltabs(~Sex + Age + Educ, df)
   ltabs(df$Sex)
   ltabs(df$Sex, df$Age)
}
}
\seealso{
\code{\link[stats]{xtabs}}, \code{\link{lfreq}}, \code{\link{lfactor}}
}

