% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects}

\item{plotlist}{a list of ggplot objects}

\item{cols}{number of columns in layout}

\item{layout}{a matrix specifying the layout. if present, \code{cols} is ignored}
}
\description{
Renders multiple ggplot plots in one image
}
\note{
If the layout is something like \code{matrix(c(1,2,3,3), nrow=2, byrow=TRUE)}, 
then plot 1 will go in the upper left, 2 will go in the upper right, and 3 will go all 
the way across the bottom.
}
\examples{
require(ggplot2)
p1 <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p2 <- ggplot(mtcars, aes(factor(cyl))) + geom_bar()
p3 <- ggplot(mtcars, aes(factor(cyl), mpg)) + geom_violin()
p4 <- ggplot(mtcars, aes(factor(cyl), mpg)) + geom_boxplot()

multiplot(p1, p2, p3, p4)
multiplot(p1, p2, p3, p4, cols = 2)
multiplot(p1, p2, p3, layout = matrix(c(1,2,3,3), nrow=2))
multiplot(p1, p2, p3, layout = matrix(c(1,2,3,3), nrow=2, byrow=TRUE))

}
\references{
Winston Chang, \emph{Cookbook for R}, 
\url{http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}
}

