% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formating.R
\name{percent}
\alias{comp.percent}
\alias{comp.percent0}
\alias{comp.percent1}
\alias{comp.percent2}
\alias{comp.percent3}
\alias{comp.percent4}
\alias{comp.percent5}
\alias{percent}
\alias{percent0}
\alias{percent1}
\alias{percent2}
\alias{percent3}
\alias{percent4}
\alias{percent5}
\title{Percent formatter (English)}
\usage{
percent(x, digits = 1)

percent0(x)

percent1(x)

percent2(x)

percent3(x)

percent4(x)

percent5(x)

comp.percent(x, digits = 1)

comp.percent0(x)

comp.percent1(x)

comp.percent2(x)

comp.percent3(x)

comp.percent4(x)

comp.percent5(x)
}
\arguments{
\item{x}{a numeric vector to format}

\item{digits}{number of decimal digits to display}
}
\description{
Multiply by 100 and display a percent sign.
}
\details{
\code{percent0(x)} is a shortcut for \code{percent(x, digits = 0)},
\code{percent1(x)} for \code{percent(x, digits = 1)}, etc.

\code{comp.percent(x)} returns the complement to 1 of \code{x},
i.e. \code{percent(1 - x)}.
}
\examples{
percent(0.004)
percent0(c(0.000045, 0.56789, 234.567))
percent2(c(0.000045, 0.56789, 234.567))
percent5(c(0.000045, 0.56789, 234.567))
comp.percent(0.255)
}

