% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formating.R
\name{pourcent}
\alias{comp.pourcent}
\alias{comp.pourcent0}
\alias{comp.pourcent1}
\alias{comp.pourcent2}
\alias{comp.pourcent3}
\alias{comp.pourcent4}
\alias{comp.pourcent5}
\alias{pourcent}
\alias{pourcent0}
\alias{pourcent1}
\alias{pourcent2}
\alias{pourcent3}
\alias{pourcent4}
\alias{pourcent5}
\title{Percent formatter (French)}
\usage{
pourcent(x, digits = 1)

pourcent0(x)

pourcent1(x)

pourcent2(x)

pourcent3(x)

pourcent4(x)

pourcent5(x)

comp.pourcent(x, digits = 1)

comp.pourcent0(x)

comp.pourcent1(x)

comp.pourcent2(x)

comp.pourcent3(x)

comp.pourcent4(x)

comp.pourcent5(x)
}
\arguments{
\item{x}{a numeric vector to format}

\item{digits}{number of decimal digits to display}
}
\description{
Multiply by 100, use a comma as decimal separator and display a percent sign (French way).
}
\details{
\code{pourcent0(x)} is a shortcut for \code{pourcent(x, digits = 0)},
\code{pourcent1(x)} for \code{pourcent(x, digits = 1)}, etc.

\code{comp.pourcent(x)} returns the complement to 1 of \code{x},
i.e. \code{pourcent(1 - x)}.
}
\examples{
pourcent(0.004)
pourcent0(c(0.000045, 0.56789, 234.567))
pourcent2(c(0.000045, 0.56789, 234.567))
pourcent5(c(0.000045, 0.56789, 234.567))
comp.pourcent(0.255)
}

