% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.ci.R
\name{prop.ci}
\alias{prop.ci}
\alias{prop.ci.lower}
\alias{prop.ci.upper}
\title{Proportion confidence interval}
\usage{
prop.ci(x, n = NULL, bounds = 1:2, ...)

prop.ci.lower(x, n = NULL, ...)

prop.ci.upper(x, n = NULL, ...)
}
\arguments{
\item{x}{a factor vector, a logical vector, a one-dimensional table or a vector of
count of successes (in that case, \code{n} should be provided)}

\item{n}{a vector of count of trials (\code{x} should be a vector of successes in that case)}

\item{bounds}{\code{1:2} for lower and upper bounds, \code{1} for lower bounds only,
\code{2} for upper bound only.}

\item{...}{additional parameters used by \code{\link[stats]{prop.test}}}
}
\value{
a vector with confidence interval
}
\description{
Provide a convenient wrapper for \code{\link[stats]{prop.test}} to compute
confidence intervals for proportions
}
\details{
\code{prop.ci.lower} is a wrapper for \code{prop.ci} with \code{bounds = 1} and
\code{prop.ci.upper} for \code{prop.ci} with \code{bounds = 2}.

When \code{x} and \code{n} are provided, they should be of same length. Alternativaly,
\code{n} could be a single integer that will be used for each element of \code{x}.

When \code{x} and \code{n} are provided, if \code{x} contains several elements,
\code{bounds} should be \code{1} or \code{2} but can't be \code{1:2}. In that scenario,
the returned vector will contain the corresponding lower or the upper confidence interval
bound for each element of \code{x}.

The confidence level could be specify with \code{conf.level} (0.95 by default).

By default, \code{NA} value are removed.
}
\examples{
if (require(questionr)) {
  data(hdv2003)
  d <- hdv2003

  freq(d$sport)
  prop.ci(d$sport)
  prop.ci.lower(d$sport)
  prop.ci.upper(d$sport)
  prop.ci(d$sport, conf.level = 0.9)
  prop.ci(table(d$sport))
  prop.ci(d$sport == "Non")
  prop.ci(d$sport == "Oui")

  prop.ci.lower(c(1277, 723), n = 2000)
  prop.ci.upper(c(1277, 723), n = 2000)

  if (require(data.table)) {
    d <- as.data.table(d)
    res <- d[, .(freq = .N), by = .(sexe, sport)]
    res[, n := sum(freq), by = sexe]
    res[, prop := freq / n]
    res[, prop.l := prop.ci.lower(freq, n)]
    res[, prop.h := prop.ci.upper(freq, n)]
    res
  }
}
}
\seealso{
\code{\link[stats]{prop.test}}
}
