% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_preserve_sum.R
\name{round_preserve_sum}
\alias{round_preserve_sum}
\title{Round values while preserve their rounded sum in R}
\source{
\url{https://www.r-bloggers.com/round-values-while-preserve-their-rounded-sum-in-r/} and
\url{http://stackoverflow.com/questions/32544646/round-vector-of-numerics-to-integer-while-preserving-their-sum}
}
\usage{
round_preserve_sum(x, digits = 0)
}
\description{
In general, the sum of rounded numbers (e.g., using the 
\code{base::round} function) is not the same as their rounded sum.
This solution applies the following algorithm
\enumerate{
   \item Round down to the specified number of decimal places
   \item Order numbers by their remainder values
   \item Increment the specified decimal place of values with ‘k’ largest remainders, 
   where ‘k’ is the number of values that must be incremented to preserve their rounded sum
}
}
\examples{
sum(c(0.333, 0.333, 0.334))
round(c(0.333, 0.333, 0.334), 2)
sum(round(c(0.333, 0.333, 0.334), 2))
round_preserve_sum(c(0.333, 0.333, 0.334), 2)
sum(round_preserve_sum(c(0.333, 0.333, 0.334), 2))
}

