% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{s.freq}
\alias{s.freq}
\title{Plot of factorial maps with frequency of each point}
\source{
\url{http://joseph.larmarange.net/?Representer-des-effectifs-dans-le}
}
\usage{
s.freq(dfxy, xax = 1, yax = 2, ...)
}
\arguments{
\item{dfxy}{a data frame with two coordinates}

\item{xax}{column for the x axis}

\item{yax}{column for the y axis}

\item{...}{additional parameters sent to \code{\link[ade4]{s.value}}}
}
\description{
For each point on the factorial plan, display the number of identical
observations (i.e. with the same coordinates).
}
\examples{
if (require(ade4) & require(questionr)) {
  data(hdv2003)
  acm <- dudi.acm(hdv2003[,c("sexe","sport","bricol","cuisine","cinema")], scannf = FALSE)
  s.label(acm$li, clabel = 0, cpoint = 2)
  s.freq(acm$li, csize = 0.75)
}
}
\seealso{
\code{\link[ade4]{s.value}}, \code{\link[ade4]{s.label}}
}
