% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{seq_makeRiver}
\alias{seq_makeRiver}
\title{Create a riverplot from a sequence object}
\usage{
seq_makeRiver(seq)
}
\arguments{
\item{seq}{a stslist object (typically produced with \code{\link[TraMineR]{seqdef}})}
}
\value{
a river plot object that can be plotted with \code{\link[riverplot]{riverplot}}
}
\description{
Create a riverplot from a sequence object
}

