% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fill_labels.R
\name{stat_fill_labels}
\alias{stat_fill_labels}
\title{Computes labels for position = "fill"}
\usage{
stat_fill_labels(mapping = NULL, data = NULL, geom = "text",
  position = "identity", width = NULL, na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
This stat makes it easier to display labels with \code{position_fill}.
}
\details{
Computed variables: \code{count}, the number of observations; \code{prop}, proportion on \code{x}; \code{cumprop}, cumulative proportion; \code{ylabel}, \code{y} position of labels.
}
\examples{
ggplot(as.data.frame(Titanic)) + 
  aes(x = Class, fill = Survived, weight = Freq) + 
  geom_bar(position = "fill") + geom_text(stat = "fill_labels")
ggplot(as.data.frame(Titanic)) + 
  aes(x = Class, fill = Survived, weight = Freq) + 
  geom_bar(position = "fill") + stat_fill_labels()
ggplot(as.data.frame(Titanic)) + 
  aes(x = Class, fill = Survived, weight = Freq) + 
  geom_bar(position = "fill") + geom_text(stat = "fill_labels") + facet_grid(~Sex)
ggplot(as.data.frame(Titanic)) + 
  aes(x = as.integer(Class), fill = Survived, weight = Freq) + 
  geom_area(position = "fill", stat = "count") + geom_text(stat = "fill_labels")
}

