% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_stack_labels.R
\name{stat_stack_labels}
\alias{stat_stack_labels}
\title{Computes labels position for position = "stack"}
\usage{
stat_stack_labels(mapping = NULL, data = NULL, geom = "text",
  position = "identity", width = NULL, na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
This stat makes it easier to display labels with \code{position_stack}.
}
\details{
Computed variables: \code{count}, the number of observations; \code{cumcount}, cumulative count; \code{ylabel}, \code{y} position of labels.
}
\examples{
ggplot(as.data.frame(Titanic)) + 
  aes(x = Class, fill = Survived, weight = Freq) + 
  geom_bar() + geom_text(stat = "stack_labels")
ggplot(as.data.frame(Titanic)) + 
  aes(x = Class, fill = Survived, weight = Freq) + 
  geom_bar() + stat_stack_labels()
ggplot(as.data.frame(Titanic)) + 
  aes(x = Class, fill = Survived, weight = Freq) + 
  geom_bar() + stat_stack_labels() + facet_grid(~Sex)
ggplot(as.data.frame(Titanic)) + 
  aes(x = as.integer(Class), fill = Survived, weight = Freq) + 
  geom_area(stat = "count") + stat_stack_labels()
}

