% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_xprop.R
\name{stat_xprop}
\alias{stat_xprop}
\title{Computes proportions by x}
\usage{
stat_xprop(mapping = NULL, data = NULL, geom = "bar",
  position = "stack", width = NULL, na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
This stat makes it easier to generate cumulative stacked bar / area charts.
}
\examples{
ggplot(as.data.frame(Titanic)) + aes(x = Class, fill = Survived, weight = Freq) + geom_bar(stat = "count")
ggplot(as.data.frame(Titanic)) + aes(x = Class, fill = Survived, weight = Freq) + geom_bar(stat = "xprop")
ggplot(as.data.frame(Titanic)) + aes(x = Class, fill = Survived, weight = Freq) + geom_bar(stat = "count") + facet_wrap(~Sex)
ggplot(as.data.frame(Titanic)) + aes(x = Class, fill = Survived, weight = Freq) + geom_bar(stat = "xprop") + facet_wrap(~Sex)
ggplot(as.data.frame(Titanic)) + aes(x = Class, fill = Survived, weight = Freq) + geom_bar(stat = "xprop", position = "dodge")
ggplot(as.data.frame(Titanic)) + aes(x = as.integer(Class), fill = Survived, weight = Freq) + geom_area(stat = "xprop")
}

