% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_local_as_cran.R
\name{test_local_as_cran}
\alias{test_local_as_cran}
\title{Test a local source package}
\usage{
test_local_as_cran(path = ".", reporter = NULL, ..., load_package = "source")
}
\arguments{
\item{path}{Path to directory containing tests.}

\item{reporter}{Reporter to use to summarise output. Can be supplied
as a string (e.g. "summary") or as an R6 object
(e.g. \code{SummaryReporter$new()}).

See \link[testthat]{Reporter} for more details and a list of built-in reporters.}

\item{...}{Additional arguments passed to \code{\link[testthat:test_dir]{test_dir()}}}

\item{load_package}{Strategy to use for load package code:
\itemize{
\item "none", the default, doesn't load the package.
\item "installed", uses \code{\link[=library]{library()}} to load an installed package.
\item "source", uses \code{\link[pkgload:load_all]{pkgload::load_all()}} to a source package.
To configure the arguments passed to \code{load_all()}, add this
field in your DESCRIPTION file:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Config/testthat/load-all: list(export_all = FALSE, helpers = FALSE)
}\if{html}{\out{</div>}}
}}
}
\description{
Similar to \code{testthat::test_local()}, it runs all tests in a package
except those skipped by \code{testthat::skip_on_cran()}, using
\code{withr::local_envvar(NOT_CRAN = "false")}.
}
