% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_msSurv.R
\name{tidy.msSurv}
\alias{tidy.msSurv}
\title{Experimental tidier for msSurv objects}
\usage{
tidy.msSurv(
  x,
  extract = c("state occupation probabilities", "transition probabilities"),
  conf.level = 0.95,
  conf.fun = c("linear", "log", "log-log", "cloglog"),
  ...
)
}
\arguments{
\item{x}{a msSurv object}

\item{extract}{what should be extracted?}

\item{conf.level}{confidence level of the intervals}

\item{conf.fun}{transformation appelied to confidence intervals}

\item{...}{extra arguments}
}
\description{
Experimental tidier for msSurv objects
}
