% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_msm.R
\name{tidy.prevalence.msm}
\alias{tidy.prevalence.msm}
\alias{tidy.hazard.msm}
\title{Basic tieders for msm models}
\usage{
\method{tidy}{prevalence.msm}(x)

\method{tidy}{hazard.msm}(x)
}
\arguments{
\item{x}{an object returned by \code{\link[msm]{prevalence.msm}} or \code{\link[msm]{hazard.msm}}}
}
\description{
Basic tieders for msm models
}
\examples{
\dontrun{
if (require(msm)) {
  twoway4.q <- rbind(
    c(-0.5, 0.25, 0, 0.25), c(0.166, -0.498, 0.166, 0.166),
    c(0, 0.25, -0.5, 0.25), c(0, 0, 0, 0)
  )
  cav.msm <- msm(state ~ years,
    subject = PTNUM, data = cav,
    qmatrix = twoway4.q, deathexact = 4, covariates = ~sex
  )
  tidy.prevalence.msm(prevalence.msm(cav.msm))
  tidy.prevalence.msm(prevalence.msm(cav.msm, ci = "normal"))
  tidy.hazard.msm(hazard.msm(cav.msm))
}
}
}
