% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_chisq.R
\name{tidy_chisq}
\alias{tidy_chisq}
\title{Tidiers for lists returned from chisq.test}
\usage{
tidy_chisq(x, ...)
}
\arguments{
\item{x}{list returned by \code{chisq.test} or \code{svychisq}}

\item{...}{extra arguments}
}
\description{
Tidies objects returned by the \code{\link[stats]{chisq.test}} function.
}
\examples{
test <- chisq.test(xtabs(Freq ~ Sex + Class, data = as.data.frame(Titanic)))
tidy_chisq(test)
}
