% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_detailed.R
\name{tidy_detailed}
\alias{tidy_detailed}
\title{Tidy data with variable names and factor levels}
\source{
This function has been adapted from Benjamin Nutter's \code{tidy_levels_labels} function
implemented within the \code{pixiedust} package
(\url{https://github.com/nutterb/pixiedust/blob/master/R/tidy_levels_labels.R}).
}
\usage{
tidy_detailed(x, ...)
}
\arguments{
\item{x}{model to tidy}

\item{...}{extra arguments passed to \code{\link[broom]{tidy}} function from \pkg{broom} package}
}
\value{
The \code{data.frame} produced by \code{\link[broom]{tidy}} with these additional columns:
\describe{
\item{variable}{variable name}
\item{variable_label}{label of the variable (if available), variable name otherwise}
\item{level}{for factors, the corresponding level}
\item{level_detail}{level with indication of the reference level, eg. \code{B vs. A}}
\item{label}{equal to level_detail if it's a factor, to variable_label otherwise}
}
}
\description{
This function extends the output of \code{\link[broom]{tidy}} function from \pkg{broom} package
by providing additional column indicating variables and factors names.
}
\note{
This function requires the \pkg{broom} and \pkg{dplyr} packages.
}
\examples{
reg <- lm(Sepal.Length ~ Sepal.Width + Petal.Length * Species, data = iris)
tidy_detailed(reg)

d <- as.data.frame(Titanic)
reg2 <- glm(Survived ~ Sex + Age + Class, family = binomial, data = d, weights = d$Freq)
tidy_detailed(reg2, conf.int = TRUE, exponentiate = TRUE)
}
\seealso{
\code{\link[broom]{tidy}} function from \pkg{broom} package
}
