% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{who_formatter}
\alias{who_formatter}
\title{Formatting numbers like in WHO publications}
\usage{
who_formatter(x)
}
\arguments{
\item{x}{a numeric vector to format}
}
\description{
In WHO publications, population numbers are usually presented as follow:
\itemize{
  \item 0 if < 5
  \item 10 if ≥ 5 and < 10
  \item one significant digit if ≥ 10 and < 100
  \item two significant digits if ≥ 100 and < 10 millions
  \item three significant digits if ≥ 10 millions
}
}
\examples{
who_formatter(c(3, 8, 42, 75, 45678, 9876543, 12345678))
}
