% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A2Rplot.R
\name{cutree.order}
\alias{cutree.order}
\title{Cut a tree into groups of data in the order of the tree}
\source{
\url{http://addictedtor.free.fr/packages/A2R/lastVersion/}
}
\usage{
cutree.order(tree, k = NULL, h = NULL)
}
\arguments{
\item{tree}{an \code{\link{hclust}} object}

\item{k}{the desired number of groups}

\item{h}{height where the tree is to be cut}
}
\description{
Cut a tree (result from \link{hclust}) into groups of data. 
Groups are in the order of the tree leafs.
}
\note{
The A2R package has not been updated since January 2006 and cannot be installed
anymore with a recent version of R. It's why this function has been copied here.
}
\examples{
h77 <- hclust(dist(state.x77))
ct.o <- cutree.order(h77, k=4)
ct.n <- cutree(h77,k=4)
plot(h77)
rect.hclust(h77, 4)
cbind(ct.o, ct.n)
}
\author{
Romain Francois
}
\seealso{
\code{\link{cutree}}
}

