% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{ggchisq_res}
\alias{ggcross}
\alias{stat_stack_labels}
\alias{stat_fill_labels}
\alias{tidy_detailed}
\title{Deprecated functions}
\usage{
ggchisq_res(...)

ggcross(...)

stat_stack_labels(...)

stat_fill_labels(...)

tidy_detailed(...)
}
\arguments{
\item{...}{parameters}
}
\description{
\lifecycle{deprecated}

These functions are deprecated.

\code{tidy.clm}, \code{tidy.clmm} and \code{tidy.svyolr} are now part of \pkg{broom}
and have been removed from \pkg{JLutils}.

\code{tidy_chisq} is deprecated. Use \code{\link[broom]{augment.htest}}
from \pkg{broom} instead.

\code{tidy_detailed} is deprecated. Use \code{\link[broom.helpers]{tidy_plus_plus}} from \pkg{broom.helpers} instead.

\code{ggchisq_res} and \code{\link{ggcross}} are now deprecated. You could use instead \code{\link[GGally]{ggtable}} from \pkg{GGally}.
}
