% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{ggchisq_res}
\title{Deprecated functions}
\usage{
ggchisq_res(...)
}
\arguments{
\item{...}{parameters}
}
\description{
These functions are deprecated.
}
\details{
\code{tidy.clm}, \code{tidy.clmm} and \code{tidy.svyolr} are now part of \pkg{broom}
and have been removed from \pkg{JLutils}.

\code{tidy_chisq} is deprecated. Use \code{\link[broom]{augment.htest}}
from \pkg{broom} instead.

\code{ggchisq_res} is now deprecated. You could use instead \code{\link[GGally]{ggtable}}
from \pkg{GGally} or \code{\link{ggcross}}.
}
