% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcross.R
\name{ggcross}
\alias{ggcross}
\title{Cross-tabulation matrix plot}
\usage{
ggcross(formula, data, weight = NULL, addNA = FALSE, fill = "residuals",
  fill_breaks = c(-4, -2, 2, 4), palette = "RdBu",
  fill_title = "Pearson residuals", size = "observed", max_size = 20,
  size_title = "Observations", labels = NULL, labels_size = 3.5,
  return_data = FALSE)
}
\arguments{
\item{formula}{formula of variables to be cross-tabulated, rows on left hand side and columns on the right hand side}

\item{data}{data frame containing the data}

\item{weight}{optionnal string indicating a column containing weights}

\item{addNA}{whether to include NA values in the tables}

\item{fill}{variable mapped to fill OR color name}

\item{fill_breaks}{how to cut fill variable into categories? (cf. \code{\link[base]{cut}})}

\item{palette}{Brewer colour palette (see \url{http://colorbrewer2.org})}

\item{fill_title}{legend title for fill}

\item{size}{variable mapped to size}

\item{max_size}{size of largest point}

\item{size_title}{legend title for size}

\item{return_data}{return computed data.frame instead of ggplot?}

\item{label}{optionnal cell labels (see examples)}

\item{label_size}{size of cell labels}
}
\value{
a ggplot graphic or a data frame if \code{return_data == TRUE}.
}
\description{
By default, plot chi-squared residuals
}
\examples{
ggcross(Sex + Age + Class ~ Survived, data = as.data.frame(Titanic), weight = "Freq")
ggcross(Sex + Age + Class ~ Survived, data = as.data.frame(Titanic), weight = "Freq", return_data = TRUE)
ggcross(Sex + Age + Class ~ Survived, data = as.data.frame(Titanic), weight = "Freq", label = "scales::percent(row.prop)")
ggcross(Sex + Age + Class ~ Survived, data = as.data.frame(Titanic), weight = "Freq", breaks = c(-4, -2, 0, 2, 4))
}

