% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/label_freq.R
\name{ltabs}
\alias{ltabs}
\title{Labelled table}
\usage{
ltabs(x, d = NULL, add.variable.labels = FALSE, ...)
}
\arguments{
\item{x}{a formula (see \code{\link[stats]{xtabs}}) or a vector}

\item{d}{a data.frame (if \code{x} is a formula) or an optional second vector}

\item{add.variable.labels}{add variable labels to dimension names?}

\item{...}{additional parameters sent to \code{\link[stats]{xtabs}}}
}
\description{
Create a contingency table (with \code{\link[stats]{xtabs}}) and add value and variable
labels if they are available (typically if the data frame has been imported with
\code{\link[foreign]{read.spss}} or \code{\link[foreign]{read.dta}})
}
\details{
This function is particulary useful when the data has been imported
from SPSS or Stata but the labelled variables not converted to factors.
}
\examples{
\dontrun{
   ltabs(~Sex, df)
   ltabs(~Sex+Age, df)
   ltabs(df$Sex)
   ltabs(df$Sex, df$Age)
}
}
\seealso{
\code{\link[stats]{xtabs}}, \code{\link{lfreq}}
}

