% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{pval}
\alias{pval}
\alias{pval_format}
\alias{signif_stars}
\title{p-values formatter and significance stars}
\usage{
pval(x, formatter = en, accuracy = 0.001, stars = FALSE,
  three = 0.001, two = 0.01, one = 0.05, point = 0.1,
  add_p = FALSE, ...)

pval_format(formatter = en, accuracy = 0.001, stars = FALSE,
  three = 0.001, two = 0.01, one = 0.05, point = 0.1,
  add_p = FALSE, ...)

signif_stars(x, three = 0.001, two = 0.01, one = 0.05, point = 0.1)
}
\arguments{
\item{x}{a numeric vector of p-values}

\item{formatter}{a formatter function, see \code{\link{number}}, typically \code{\link{en}} or \code{\link{fr}}}

\item{accuracy}{number to round to}

\item{stars}{add significance stars?}

\item{three}{level below which to display three stars '***'}

\item{two}{level below which to display two stars '**'}

\item{one}{level below which to display one star '*'}

\item{point}{level below which to display a point '.' (\code{NULL} to not display a point)}

\item{add_p}{add "p=" before the value?}

\item{...}{additional parameters sent to formatter function}
}
\description{
Formatter for p-values, adding a symbol "<" for very small p-values and, optionally, significance stars
}
\details{
\code{pval_format} will produce a custom function, to be used for example with \code{ggplot2}.
}
\examples{
p <- c(.50, 0.12, .09, .045, .011, .009, .00002, NA)
pval(p)
pval(p, formatter = fr, accuracy = .01)
pval(p, stars = TRUE)
pval(p, add_p = TRUE)
custom_function <- pval_format(accuracy = .1, stars = TRUE)
custom_function(p)
signif_stars(p)
signif_stars(p, one = .15, point = NULL)
}
