% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{pvalue}
\alias{pvalue}
\alias{label_pvalue}
\alias{signif_stars}
\title{p-values formatter}
\usage{
pvalue(x, accuracy = 0.001, decimal.mark = ".", prefix = NULL, add_p = FALSE)

label_pvalue(
  accuracy = 0.001,
  decimal.mark = ".",
  prefix = NULL,
  add_p = FALSE
)

signif_stars(x, three = 0.001, two = 0.01, one = 0.05, point = 0.1)
}
\arguments{
\item{x}{A numeric vector to format.}

\item{accuracy}{A number to round to. Use (e.g.) \code{0.01} to show 2 decimal
places of precision. If \code{NULL}, the default, uses a heuristic that should
ensure breaks have the minimum number of digits needed to show the
difference between adjacent values.

Applied to rescaled data.}

\item{decimal.mark}{The character to be used to indicate the numeric
decimal point.}

\item{prefix}{A character vector of length 3 giving the prefixes to
put in front of numbers. The default values are \code{c("<", "", ">")}
if \code{add_p} is \code{TRUE} and \code{c("p<", "p=", "p>")} if \code{FALSE}.}

\item{add_p}{Add "p=" before the value?}

\item{three}{threshold below which to display three stars}

\item{two}{threshold below which to display two stars}

\item{one}{threshold below which to display one star}

\item{point}{threshold below which to display one point (\code{NULL} to desactivate)}
}
\value{
\code{label_pvalue} returns a function with single parameter
\code{x}, a numeric vector, that returns a character vector.
}
\description{
Formatter for p-values, adding a symbol "<" for small p-values.
}
\examples{
p <- c(.50, 0.12, .045, .011, .009, .00002, NA)
pvalue(p)
pvalue(p, accuracy = .01)
pvalue(p, add_p = TRUE)
custom_function <- label_pvalue(accuracy = .1, decimal.mark = ",")
custom_function(p)
signif_stars(p)
signif_stars(p, one = .15, point = NULL)
}
