% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{pvalue}
\alias{pvalue}
\alias{pvalue_format}
\alias{signif_stars}
\title{p-values formatter}
\usage{
pvalue(x, accuracy = 0.001, decimal.mark = ".", add_p = FALSE)

pvalue_format(accuracy = 0.001, decimal.mark = ".", add_p = FALSE)

signif_stars(x, three = 0.001, two = 0.01, one = 0.05, point = 0.1)
}
\arguments{
\item{x}{A numeric vector of p-values.}

\item{accuracy}{Number to round to.}

\item{decimal.mark}{The character to be used to indicate the numeric
decimal point.}

\item{add_p}{Add "p=" before the value?}

\item{three}{threshold below which to display three stars}

\item{two}{threshold below which to display two stars}

\item{one}{threshold below which to display one star}

\item{point}{threshold below which to display one point (\code{NULL} to desactivate)}
}
\value{
`pvalue_format` returns a function with single parameter
  `x`, a numeric vector, that returns a character vector.
}
\description{
Formatter for p-values, adding a symbol "<" for small p-values.
}
\examples{
p <- c(.50, 0.12, .045, .011, .009, .00002, NA)
pvalue(p)
pvalue(p, accuracy = .01)
pvalue(p, add_p = TRUE)
custom_function <- pvalue_format(accuracy = .1, decimal.mark = ",")
custom_function(p)
signif_stars(p)
signif_stars(p, one = .15, point = NULL)
}
