% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{stat_prop}
\alias{stat_weighted_mean}
\alias{stat_cross}
\alias{ggsurvey}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{GGally}{\code{\link[GGally]{stat_cross}}, \code{\link[GGally]{stat_prop}}, \code{\link[GGally]{stat_weighted_mean}}}

  \item{questionr}{\code{\link[questionr]{ggsurvey}}}
}}

