\name{seq.heatmap}
\alias{seq.heatmap}
\title{Index plot of sequences oderred according to a dendrogram}
\source{
  \url{http://joseph.larmarange.net/?Representer-un-tapis-de-sequences}
}
\usage{
  seq.heatmap(seq, tree, with.missing = FALSE, ...)
}
\description{
  Index plot of sequences oderred according to a dendrogram
}

